Add showRedundant flag.
To illustrate the actual code dependencies in a project (especially when
enforcing total dependency definition using e.g. the dependency plugin),
it is necessary to have all edges from projects to dependencies present.

Currently, in a dependency tree like this:

                A
               / \
              B   C
             /
            C

the graph plugin will only plot

               A
               |
               B
               |
               C

This pull request adds a flag that will not remove any "redundant"
edges, so the result will be

               A
              / \
             B   |
              \ /
               C

While these graphs are more cluttered, they are also very useful to
fully grok dependencies in large Maven projects.